requirements.txt
transformers
bitsandbytes
accelerate
pillow
--index-url https://download.pytorch.org/whl/cu118
torch==2.1.0 
--index-url https://download.pytorch.org/whl/cu118
torchvision==0.16.0
--index-url https://download.pytorch.org/whl/cu118
torchaudio==2.1.0


